test(BA-6725): merge enqueue delegation authz into action test file
No need for a separate file — fold the mocked authorization tests into
test_enqueue_session_action.py alongside the scope-targeting tests, both
covering EnqueueSessionAction owner delegation.
